refactor: subways 데이터가 있으면 실행하는 조건 추가

+++ src/main/java/com/knu/subway/service/SubwayDataCollector.java
@@ -66,12 +66,13 @@ public class SubwayDataCollector {
 
         // 5분 전보다 업데이트된 데이터 조회
         List<Subway> subways = subwayService.findByUpdatedIsBefore(fiveMinutesAgo);
-
-        for (Subway subway : subways) {
-            // 데이터가 5분 이상 된 경우 처리
-            log.info("Processing old data for subway: {}", subway.getId());
-            subwayService.delete(subway);
-            subwayCookie.add(subway.getBtrainNo());
+        if(!subways.isEmpty()) {
+            for (Subway subway : subways) {
+                // 데이터가 5분 이상 된 경우 처리
+                log.info("Processing old data for subway: {}", subway.getId());
+                subwayService.delete(subway);
+                subwayCookie.add(subway.getBtrainNo());
+            }
         }
     }
     @Scheduled(cron = "0 0 5 * * *")
